+++ lib/time_machine/validation/validator_factory.rb
@@ -14,6 +14,7 @@ require './lib/time_machine/validators/geom_new_object'
 require './lib/time_machine/validators/network'
 require './lib/time_machine/validators/tags_changes'
 require './lib/time_machine/validators/user_block'
+require './lib/time_machine/validators/user_new'
 require './lib/time_machine/validators/user_list'
 require './lib/time_machine/validators/validator_link'