+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Clearance Changelog
 
 * Add validator `user_block`: change made by an user currently blocked, or with too much previous blocks.
+* Add validator `user_new`: change made by an user with only few changesets, or user created only for few days ago.
 
 ## v0.6.0 (2026-08-20)